requirements.txt
discord.py[voice] >= 2.2.0
ipython
discord-racket >= 0.0.13